ボックスの投影を完成させました。 またそれに付随して衝突時関数の二次元版の作成。 衝突イベントシステムのコライダー削除を追加しました。

+++ Projection/GameLib/GameObject/Components/Collider/2D/Shapes/ConvexPolygonCollider2D.cpp
@@ -67,7 +67,7 @@ void ConvexPolygonCollider2D::UpdateCache() const
     }
 
     // AABB‚ÌXV
-    if (int count = m_worldVertices.size(); count > 0)
+    if (size_t count = m_worldVertices.size(); count > 0)
     {
         SimpleMath::Vector2 min = m_worldVertices[0], max = m_worldVertices[0];
